ai-chatbot/ai-chatbot/pyproject.toml
[project]
name = "stratix-ai-chat-bot"
version = "1.0.0"
description = "Multi-channel AI chatbot system for phone, Telegram, Instagram, and web"
authors = [{ name = "STRATIX Team", email = "dev@stratix.uz" }]
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "pyjwt>=2.8.0",
    "httpx>=0.26.0",
    "dishka>=1.0.0",
    "celery>=5.3.6",
    "structlog>=25.5.0",
    "starlette>=0.52.1",
    "greenlet>=3.3.2",
    "uvloop>=0.22.1",
    "alembic>=1.18.4",
    "celery-types>=0.24.0",
    "gunicorn>=25.1.0",
    "aiogram>=3.25.0",
    "websockets>=16.0",
    "openai>=2.21.0",
    "audioop-lts>=0.2.2",
    "retailcrm>=5.3.0",
    "typer>=0.24.1",
    "openpyxl>=3.1.0",
    "google-genai>=1.65.0",
    "langgraph>=1.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.5.1",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
    "types-python-dateutil>=2.8.19",
    "pre-commit>=3.6.0",
]

# Note: CLI entry points must be valid Python entrypoint references
# Use: uvicorn src.main:app --host 0.0.0.0 --port 8000 to start the server

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501", "B008", "B904", "ARG001"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]
"alembic/*" = ["ARG001"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["src/workers/*", "src/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
